Fix: on sad path, partial import failed due to mal-indexing
Redcap's import requires that, when using DataFrames, `api.def_field`
and,  if present, `redcap_event_name` be made into indices. In this
case, the indexing was only done on the happy path, and exception
typically happened before it.

What's notable is that the API response was completely misleading,
  suggesting an altogether wrong culprit:

```
Traceback (most recent call last):
    File "/var/www/html/applications/fitbit/ingest_api_to_redcap.py", line 165, in <module>
        out = rc_api.import_records(join, overwrite='overwrite', return_content='ids')
      File "/var/www/html/applications/fitbit/env/local/lib/python2.7/site-packages/redcap/project.py", line 487, in import_records
        raise RedcapError(str(response))
    RequestException: {u'error': u'The record id field (id_redcap) is missing.'}
```

I'm noting it in this commit in case some poor soul should happen upon
the same error, and Google it, and find nothing save this.