Wrap per-site code in try-except and log uncaught exceptions
(The change is minimal, although it may not appear so on a regular git
 diff; I recommend using the `--ignore-space-change` flag for inspecting
 this commit.)

This is motivated by a service outage between Dec 20 and Dec ~25, when a
problem with incoming data from one of the sites prevented the
processing of all other sites that followed it.

The changed code will now automatically move to the next site and log
the error in the log file with a full traceback.